- refactor utils into array and string utils
